sih25/API_ENHANCED/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
geoalchemy2==0.15.2
alembic==1.13.2
python-multipart==0.0.9
pandas==2.2.3
pyarrow==17.0.0
xarray==2024.6.0
netCDF4==1.7.1
numpy==1.26.4
scipy==1.14.1
rasterio==1.4.1
shapely==2.0.6
httpx==0.27.2
requests==2.32.3
ftputil==5.1.0
retrying==1.3.4
rich==13.8.1
sentence-transformers==3.1.1
chromadb==0.5.11
faiss-cpu==1.8.0.post1
python-dotenv==1.0.1
websockets==12.0
